app/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
# torch==1.13.1
torch==1.9.0
opencv-python-headless==4.8.1.78
opencv-python==4.8.1.78
numpy==1.20.3
torchvision==0.10.0
uvicorn==0.22.0
fastapi==0.95.2
pytest==7.3.1
starlette-prometheus==0.9.0
python-multipart==0.0.6
# python -m pip install 'git+https://github.com/facebookresearch/detectron2.git'